extract/requirements.txt
requests==2.31.0
google-cloud-secret-manager==2.16.2
google-cloud-storage==2.10.0